Move QueryData to Query.h
