+++ src/main/common/entities/ClientMessage.h
@@ -3,7 +3,6 @@
 
 #include "Message.h"
 #include "Query.h"
-#include "QueryData.h"
 
 typedef struct {
     long mtype;
