+++ src/main/common/entities/Query.h
@@ -2,8 +2,25 @@
 #define DATABASE_QUERY_H
 
 #include <string>
+#include <stddef.h>
 
-#include "QueryData.h"
+enum QUERY_TYPE {
+    SELECT,
+    INSERT
+};
+
+enum QUERY_BUFF_SIZE {
+    NAME = 62,
+    ADDRESS = 121,
+    PHONE = 14
+};
+
+typedef struct {
+    int operation;
+    char name[QUERY_BUFF_SIZE::NAME];
+    char address[QUERY_BUFF_SIZE::ADDRESS];
+    char phone[QUERY_BUFF_SIZE::PHONE];
+} QueryData;
 
 class Query {
